requirements.txt
asgiref==3.7.2
Django==3.2
django-ckeditor==6.7.0
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
phonenumbers==8.13.29
phonenumberslite==8.13.29
pillow==10.2.0
PyJWT==1.7.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
